pyproject.toml
[project]
name = "memory-service"
version = "0.1.0"
description = "Multi-tenant memory service with OpenRouter, TEI, and Qdrant backends"
requires-python = ">=3.12"
keywords = ["memory", "qdrant", "mcp", "fastapi"]
license = {text = "MIT"}
authors = [{name = "Memory Service Team"}]
readme = "README.md"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "qdrant-client>=1.9",
    "python-dotenv>=1.0",
    "orjson>=3.10",
    "uvloop>=0.19",
    "sse-starlette>=1.6",
    "cachetools>=5.3",
    "tenacity>=8.3",
    "PyJWT>=2.8",
    "cryptography>=42.0",
    "rich>=13.7",
    "pyyaml>=6.0",
    "mcp>=0.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx[http2]>=0.27",
    "respx>=0.20",
    "ruff>=0.5",
    "coverage[toml]>=7.5"
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --disable-warnings"
testpaths = ["tests"]
asyncio_mode = "auto"
